Move CartItem to features
