+++ apps/store/src/app/[locale]/confirmation/[shopSessionId]/ConfirmationPage.tsx
@@ -5,9 +5,9 @@ import { Heading, mq, Space, theme } from 'ui'
 import * as GridLayout from '@/components/GridLayout/GridLayout'
 import { ShopBreakdown } from '@/components/ShopBreakdown/ShopBreakdown'
 import { ProductItemContractContainerCar } from '@/features/carDealership/ProductItemContractContainer'
+import { CartItem } from '@/features/CartItem/CartItem'
 import { CartTotal } from '@/features/CartTotal/CartTotal'
 import { SasEurobonusSectionContainer } from '@/features/sas/SasEurobonusSection'
-import { CartItem } from '../../checkout/components/CartItem/CartItem'
 import { StaticContent } from './components/StaticContent/StaticContent'
 import { SuccessAnimation } from './components/SuccessAnimation/SuccessAnimation'
 import { SwitchingAssistantSection } from './components/SwitchingAssistantSection/SwitchingAssistantSection'
