+++ apps/store/src/components/ProductItem/EditActionButton.tsx
@@ -1,8 +1,8 @@
 import styled from '@emotion/styled'
 import { useTranslation } from 'next-i18next'
 import { Button, Text } from 'ui'
-import { useEditProductOffer } from '@/app/[locale]/checkout/components/CartItem/hooks/useEditProductOffer'
 import * as FullscreenDialog from '@/components/FullscreenDialog/FullscreenDialog'
+import { useEditProductOffer } from '@/features/CartItem/hooks/useEditProductOffer'
 import type { ProductOfferFragment } from '@/services/graphql/generated'
 import { ActionButton } from './ProductItem'
