+++ apps/store/src/features/CartItem/components/CartItemProductDetails.tsx
